Don't send UpdatedDateUTC back to Xero.
Some fields (especially this one) will cause a server error
if you attempt to send them back: it's not really clear from
the server what the error is, and the only way to resolve it is
to remove the field before sending.

This resolves #10.

It is possible this affects other fields, but at this stage I have
not noticed any.