Fix XeroBadRequest without 'Elements'
There are circumstances where the returned data may not have an 'Elements' key.  In such cases, the upstream code bombs out and fails to report any useful error detail, see freakyboy3742/pyxero#78.

This commit catches any KeyError in the problematic code expression, and stores the entire 'Message' value in the errors list as a fallback.