Catch KeyErrors in _parse_api_response
The response to an Attachment request is not structured in the same way as other requests that pass through Manager.  In particular, it does not have a key matching 'resource_name' at its top level.  This causes _parse_api_response to blow up unnecessarily.

It may be possible to do this in a more targeted way, by detecting the structure of an Attachment response, and unpacking it as appropriate, but this is a) better than nothing, and b) will also work for any other unorthodox response types that happen along in future.

Fixes freakboy3742/pyxero#94