Include error detail for validation error XeroBadRequest.
For validation errors, the value provided in 'Message' is not
actually very helpful.  The important information is in the
'ValidationErrors' values.  To make the main exception message more
useful, I have appended the exception 'problem' to the main message, for
JSON HTTP 400 responses.

See https://github.com/freakboy3742/pyxero/issues/129